backend_rest/requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.2.1
django-filter==2.2.0
django-multiselectfield==0.1.11
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
oauthlib==3.1.0
openpyxl==3.0.3
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.22.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.7
zope.interface==4.7.2
